feat: consolidate Channels and Connect settings tabs
Move Telegram and SMS channel cards from the separate Channels tab into
the Connect tab, positioned between Bearer Token and QR Pairing sections.
Remove the Channels tab from the settings nav entirely.
